Combine codec files
