+++ api/clients/config.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"time"
 
-	"github.com/Layr-Labs/eigenda/api/clients/codecs"
+	"github.com/Layr-Labs/eigenda/encoding/utils/codec"
 )
 
 type EigenDAClientConfig struct {
@@ -64,7 +64,7 @@ type EigenDAClientConfig struct {
 	DisableTLS bool
 
 	// The blob encoding version to use when writing blobs from the high level interface.
-	PutBlobEncodingVersion codecs.BlobEncodingVersion
+	PutBlobEncodingVersion codec.BlobEncodingVersion
 
 	// Point verification mode does an IFFT on data before it is written, and does an FFT on data after it is read.
 	// This makes it possible to open points on the KZG commitment to prove that the field elements correspond to
