+++ encoding/utils/codec/fft_test.go
@@ -1,4 +1,4 @@
-package codecs
+package codec
 
 import (
 	"bytes"
